pyproject.toml
[project]
name = "codebolt"
version = "0.1.0"
description = "Codebolt Application"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.2",
    "websockets>=14.2",
]
